Replace CompanyLocalHeader with refactored version
